Implement authorization using Gate

+++ app/Http/Controllers/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -55,7 +56,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        //
+        if (Gate::denies('update-question',$question)){
+            abort(403,"Access Denied");
+        }
         return view('questions.edit',['question'=>$question]);
     }
 
@@ -65,6 +68,9 @@ class QuestionController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        if (Gate::denies('update-question',$question)){
+            abort(403,"Access Denied");
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with(['success'=>"Your question has been updated success"]);
     }
@@ -75,6 +81,9 @@ class QuestionController extends Controller
     public function destroy(Question $question)
     {
         //
+        if (Gate::denies('delete-question',$question)){
+            abort(403,"Access Denied");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with(['success'=>"Your question has been deleted"]);
     }
